+++ NEWS.md
@@ -1,4 +1,5 @@
 ### largeVis 0.1.10.1
+* Hotfix for issue the caused largeVis to fail if compiled without 64-bit ARMA  
 * Fix for issue reported by @Dalar where sparse neighbor search could fail with division by zero error
 
 ### largeVis 0.1.10